Post feed

+++ CreatePostDB.sql
@@ -4,5 +4,6 @@ create table Project353.Posts(
 
 USERID    VARCHAR(25),
 content    VARCHAR(256),
-datePosted		INTEGER
+datePosted		BIGINT
+
 );
